If there is one alt and it has a constructor that is different to the LHS of the rule, generate a concrete subclass for it
Previously if there was one alt we always used the LHS name, ignoring any specified constructor name. It's useful
to be able to vary these because the LHS name is used in error messages from Rats! so you want it to be long
and descriptive, e.g, FunctionDefinition so you get "function definition" in the messages, but in the tree it's a pain
to have long constructor names, so you want to specify a shorter name, e.g, FunDef.

Some additional tidy up of the code to generate the classes.