Separate Symbol definition into SymbolN for each unique length N of symbol literals
This fixes a bug whereby having two symbols such as ?]? and ?]]? would not work. Since there
was only one Symbol definition, whichever of the two literal forms was parsed first would win,
thereby meaning that the other form could never be recognised. E.g., if ?]]? took precedence,
then an input with ]] in it could never be recognised as two ?]? literals.

The solution is to separate the symbol definitions by the length of the literals that they match
so there can be no confusion between symbols.